requirements.txt
Pillow
pyzbar
base45
wheel
cose

# Uncomment if you want to use the notebook
# jupyter

# Uncomment if you want to use the screenshot feature
# pyautogui

# Uncomment if you want to use signature verification
# requests
# cryptography